% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{add_metadata}
\alias{add_metadata}
\title{Add metadata of variables and samples.}
\usage{
add_metadata(sce = NULL, mitochondria_symbol = NULL)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{mitochondria_symbol}{A string representing for mitochondrial genes.
This function computes percents of reads that map to the mitochondrial
genes. Examples are `^MT-`, `^mt-`, etc.}
}
\value{
A SingleCellExperiment object.
}
\description{
This function adds metadata of variables and samples.
}
\examples{
data(pbmc_eg)
pbmc <- add_metadata(sce = pbmc_eg, mitochondria_symbol = "^MT-")

}
