% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASURATBI.R
\docType{package}
\name{ASURATBI}
\alias{ASURATBI}
\title{A pipeline especially built for computing several single-cell datasets.}
\description{
ASURATBI is a computational pipeline especially built for computing
   single-cell and spatial transcriptome datasets for our paper publications.
   ASURATBI is a developed version of ASURAT, using several informal packages
   and additional functions.
}
