% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bubble_sort}
\alias{bubble_sort}
\title{Perform bubble sorting, counting the number of steps.}
\usage{
bubble_sort(listdata)
}
\arguments{
\item{listdata}{A list of vector and integer.
For example, in R code, listdata = list(vec = c(1, 0, 1, ...), cnt = 0).
The integer (cnt = 0) is the initial number of steps for bubble sorting.}
}
\value{
A List.
}
\description{
Perform bubble sorting, counting the number of steps.
}
\examples{
bubble_sort(list(vec = c(1, 1, 0), cnt = 0))

}
