% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_COMSig_eg}
\alias{human_COMSig_eg}
\title{A list of small Cell Ontology and MSigDB databases for human.}
\format{
A list of dataframe.
}
\usage{
human_COMSig_eg
}
\description{
A list of small Cell Ontology and MSigDB databases for human.
}
\keyword{datasets}
