% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_KEGG_eg}
\alias{human_KEGG_eg}
\title{A list of small KEGG database for human.}
\format{
A list of dataframe.
}
\usage{
human_KEGG_eg
}
\description{
A list of small KEGG database for human.
}
\keyword{datasets}
