% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_eg}
\alias{pbmc_eg}
\title{A SingleCellExperiment object made from a gene expression table.}
\format{
SingleCellExperiment object.
}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression}
}
\usage{
pbmc_eg
}
\description{
A SingleCellExperiment object, including 50 genes and 50 cells.
  The original data "4k PBMCs from a Healthy Donor" was downloaded from 10x
  Genomics database.
}
\keyword{datasets}
