% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmcs_eg}
\alias{pbmcs_eg}
\title{A list of SingleCellExperiment objects made from sign-sample matrices.}
\format{
A list of SingleCellExperiment objects.
}
\usage{
pbmcs_eg
}
\description{
A list of SingleCellExperiment objects, consisting of small sign-by-sample
  matrices, pbmcs_eg$CM (using Cell Ontology and MSigDB databases),
  pbmcs_eg$GO (using Gene Ontology database), and pbmcs_eg$KG (KEGG).
  Here, pbmcs_eg$CM, pbmcs_eg$GO, and pbmcs_eg$KG include 87, 72, and 64
  signs, respectively, and 50 cells.
}
\keyword{datasets}
