% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dataframe2D}
\alias{plot_dataframe2D}
\title{Visualize a two-dimensional data with labels and colors.}
\usage{
plot_dataframe2D(dataframe2D = NULL, labels = NULL, colors = NULL)
}
\arguments{
\item{dataframe2D}{A dataframe with two columns.}

\item{labels}{NULL or a vector of labels of all the samples,
corresponding to colors.}

\item{colors}{NULL or a vector of colors of all the samples,
corresponding to labels.}
}
\value{
A ggplot object.
}
\description{
This function visualizes a two-dimensional data with labels and colors.
}
\examples{
data(pbmcs_eg)
labels <- SummarizedExperiment::colData(pbmcs_eg$CM)$seurat_clusters
mat <- SingleCellExperiment::reducedDim(pbmcs_eg$CM, "TSNE")
dataframe2D <- as.data.frame(mat)
plot_dataframe2D(dataframe2D = dataframe2D, labels = labels, colors = NULL)

}
