% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_additional.R
\name{plot_pseudotime3D}
\alias{plot_pseudotime3D}
\title{Plot pseudotime computed by MERLoT.}
\usage{
plot_pseudotime3D(
  ElasticTree,
  Pseudotimes,
  labels,
  colors,
  theta = 30,
  phi = 30,
  title = "",
  xlabel = "",
  ylabel = "",
  zlabel = ""
)
}
\arguments{
\item{ElasticTree}{An elastic tree object computed by MERLoT.}

\item{Pseudotimes}{A pseudotime (high-dim) object computed by MERLoT.}

\item{labels}{NULL or a vector of labels of all the samples,
corresponding to colors.}

\item{colors}{NULL or a vector of colors of all the samples,
corresponding to labels.}

\item{theta}{Angle of the plot.}

\item{phi}{Angle of the plot.}

\item{title}{Title.}

\item{xlabel}{x-axis label.}

\item{ylabel}{y-axis label.}

\item{zlabel}{z-axis label.}
}
\value{
A scatter3D object in plot3D package.
}
\description{
This function shows pseudotimes in low-dimensional spaces.
}
