% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_additional.R
\name{plot_pseudotimecourse_woTree}
\alias{plot_pseudotimecourse_woTree}
\title{Plot pseudotime courses of sign scores without elastic tree.}
\usage{
plot_pseudotimecourse_woTree(sce, signID, Pseudotimes, range_y)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{signID}{A signID.}

\item{Pseudotimes}{A pseudotime (high-dim) object computed by MERLoT.}

\item{range_y}{"cells" or "tree", which determines the range of y-axis.}
}
\value{
A ggplot object.
}
\description{
This function shows sign scores along pseudotimes.
}
