% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_additional.R
\name{plot_pseudotimecourse_wTree}
\alias{plot_pseudotimecourse_wTree}
\title{Plot pseudotime courses of sign scores with elastic tree.}
\usage{
plot_pseudotimecourse_wTree(
  sce,
  signID,
  ElasticTree,
  SignsSpaceEmbedding,
  Pseudotimes,
  labels,
  range_y
)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{signID}{A signID.}

\item{ElasticTree}{An elastic tree object computed by MERLoT.}

\item{SignsSpaceEmbedding}{A sign space embedding object computed by MERLoT.}

\item{Pseudotimes}{A pseudotime (high-dim) object computed by MERLoT.}

\item{labels}{NULL or a vector of labels of all the samples.}

\item{range_y}{"cells" or "tree", which determines the range of y-axis.}
}
\value{
A ggplot object.
}
\description{
This function shows sign scores and elastic trees along pseudotimes.
}
