% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_violin}
\alias{plot_violin}
\title{Violin plots of a one-dimensional data with labels and colors.}
\usage{
plot_violin(dataframe1D = NULL, labels = NULL, colors = NULL)
}
\arguments{
\item{dataframe1D}{A dataframe with one column.}

\item{labels}{NULL or a vector of labels of all the samples,
corresponding to colors.}

\item{colors}{NULL or a vector of colors of all the samples,
corresponding to labels.}
}
\value{
A ggplot object.
}
\description{
This function outputs violin plots of a one-dimensional data with labels
  and colors.
}
\examples{
data(pbmcs_eg)
vname <- "MSigID:92-S"
pbmc_sub <- pbmcs_eg$CM[rownames(pbmcs_eg$CM) \%in\% vname, ]
labels <- SummarizedExperiment::colData(pbmc_sub)$seurat_clusters
mat <- t(as.matrix(SummarizedExperiment::assay(pbmc_sub, "counts")))
dataframe1D <- as.data.frame(mat)
plot_violin(dataframe1D = dataframe1D, labels = labels, colors = NULL)

}
