% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swap_pass}
\alias{swap_pass}
\title{Perform one-shot adjacent swapping for each element.}
\usage{
swap_pass(listdata)
}
\arguments{
\item{listdata}{A list of vector and integer.}
}
\value{
A List.
}
\description{
Perform one-shot adjacent swapping for each element.
}
\examples{
swap_pass(list(vec = c(1, 1, 0), cnt = 0))

}
